Internship report / Master thesis
Submitted on 6th June 2024. This report presents a comprehensive account of the five-month internship project undertaken at the Helmholtz Munich under the supervision of Dr. Igor Tetko, serving as detailed documentation of the work completed during this period. It is a requisite for obtaining my MSc. in Chemoinformatics and Physical Chemistry degree within the Erasmus Mundus Joint Master's program of the Università degli Studi di Milano and Université de Strasbourg.